pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "endfield_ocr_core"
version = "1.0.0"
description = "OCR core for Endfield Stocks and Endfield Stocks Bot"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = ["endfield", "ocr"]
authors = [{ name = "Zexsch "}]
dependencies = [
  "pytesseract",
  "pillow",
  "toml"
]

[tool.hatch.build.targets.wheel]
packages = ["src/endfield_ocr_core"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "README.md",
    "LICENSE",
    "pyproject.toml"
]